Extracted components from HandPosePreset
