+++ src/js/shot-generator-new/utils/clampInstance.js
@@ -1,0 +1,13 @@
+import * as THREE from 'three'
+
+const s = new THREE.Vector3(0, 0, -1)
+const clampInstance = (instance, camera ) => {
+    let box = new THREE.Box3().setFromObject(instance)
+    let sphere = new THREE.Sphere()
+    box.getBoundingSphere(sphere)
+    let h = sphere.radius / Math.tan( camera.fov / 2 * Math.PI / 180 )
+    camera.position.addVectors( sphere.center, s.setLength(h))
+    camera.lookAt(sphere.center)
+    camera.updateMatrixWorld(true)
+}
+export default clampInstance
